reorganize and fix tests
