+++ product-service/jest.config.js
@@ -4,6 +4,6 @@ module.exports = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
-  testMatch: ['<rootDir>/src/**/*.test.(ts|tsx)'],
+  testMatch: ['<rootDir>/test/*.test.(ts|tsx)'],
   roots: ['<rootDir>'],
 };
